[FREELDR] Rename the `ext2.*` to `ext.*`
